change (migration-sale_items) quantity is numeric now and need migrate:fresh --seed

+++ database/migrations/2026_06_05_125321_create_sale_items_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('product_name');
             $table->string('product_sku')->nullable();
 
-            $table->integer('quantity');
+            $table->numeric('quantity');
             $table->enum('unit', ['kg', 'l', 'pcs']);
 
             $table->unsignedBigInteger('price');
